requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.6
pydantic[email]==2.10.6
python-dotenv==1.0.1
Jinja2==3.1.5
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
yfinance==0.2.52
httpx==0.28.1
python-jose[cryptography]==3.3.0
